+++ R/use_badges.R
@@ -42,6 +42,36 @@ use_badge_travis <- function(repo = NULL, user = "KWB-R")
   )
 }
 
+# use_badge_ghactions ----------------------------------------------------------
+
+#' Badge Github Actions
+#' @param repo name of repository (default: NULL)
+#' @param user user name or organisation under which repository defined in
+#' parameter "repo" is hosted (default: KWB-R")
+#' @return generates travis badge link
+#' @export
+use_badge_ghactions <- function(repo = NULL, user = "KWB-R")
+{
+  sprintf("%s\n%s",
+
+  kwb.utils::resolve(
+    "url",
+    grammars$general,
+    grammars$ghactions_rcmdcheck,
+    user = user,
+    pkgname = get_pkgname(repo)
+  ),
+  kwb.utils::resolve(
+    "url",
+    grammars$general,
+    grammars$ghactions_pkgdown,
+    user = user,
+    pkgname = get_pkgname(repo)
+  )
+  )
+
+}
+
 # use_badge_codecov ------------------------------------------------------------
 
 #' Badge codecov