+++ R/use_readme_md.R
@@ -28,8 +28,7 @@ use_readme_md <- function(
   docu_dev <- sprintf("%s/dev", docu_release)
 
   readme_md <- c(
-    use_badge_appveyor(pkg$name, user,domain),
-    use_badge_travis(pkg$name, user),
+    use_badge_ghactions(pkg$name, user),
     use_badge_codecov(pkg$name, user, domain ),
     use_badge_lifecycle(stage),
     use_badge_cran(pkg$name),